+++ test/muuntaja/core_perf_test.clj
@@ -439,7 +439,7 @@
   ; 4.6µs (content-type)
   ; 4.4µs && 7.1µs
   ; 3.5µs && 7.0µs (streaming)
-  (let [{:keys [enter leave]} (interceptor/format
+  (let [{:keys [enter leave]} (interceptor/format-interceptor
                                 (assoc m/default-options :return :output-stream))
         app (fn [ctx] (-> ctx enter (handle +handler+) leave :response))
         request! (context-stream +json-request+)]
@@ -453,7 +453,7 @@
   ; 8.7µs (negotiations)
   ; 8.5µs (content-type)
   ; 4.7µs && 12.4µs (streaming)
-  (let [{:keys [enter leave]} (interceptor/format
+  (let [{:keys [enter leave]} (interceptor/format-interceptor
                                 (assoc m/default-options :return :output-stream))
         app (fn [ctx] (-> ctx enter (handle +handler+) leave :response))
         request! (context-stream +transit-json-request+)]
@@ -697,7 +697,7 @@
       ;   23µs (1k)
       ;  220µs (10k)
       ; 1990µs (100k)
-      (let [{:keys [enter leave]} (interceptor/format
+      (let [{:keys [enter leave]} (interceptor/format-interceptor
                                     (assoc m/default-options :return :output-stream))
             handler (fn [ctx] (assoc ctx :response {:status 200 :body (-> ctx :request :body-params)}))
             app (fn [ctx] (-> ctx enter handler leave :response))]
@@ -708,7 +708,7 @@
       ;   16µs (1k)
       ;  153µs (10k)
       ; 1410µs (100k)
-      (let [{:keys [enter leave]} (interceptor/format
+      (let [{:keys [enter leave]} (interceptor/format-interceptor
                                     (assoc m/default-options :return :output-stream))
             handler (fn [ctx] (assoc ctx :response {:status 200 :body (-> ctx :request :body-params)}))
             app (fn [ctx] (-> ctx enter handler leave :response))]